requirements.txt
streamlit==1.32.0
python-docx==1.1.0 